Optimized memory usage of simple_filter_amplicon_mk
List of reads is no more stored in memory. We used generator instead.
Number of reads is read from bai index.
There was a bug in two - way amplicon detection so we drop it for now and we get back to one way amplicon selection.
Time of every cycle is still way to long....